refact(Collapsible): abastract radix collapsible animation into a shared component: Collapsible
