+++ apps/store/src/components/ProductItem/ProductItem.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled'
 import { useTranslation } from 'next-i18next'
 import React, { type ComponentProps, useState, forwardRef, type ReactNode } from 'react'
-import type { ButtonProps} from 'ui';
+import type { ButtonProps } from 'ui'
 import { Badge, Button, Text, mq, theme } from 'ui'
-import * as Collapsible from '@/components/Collapsible'
+import Collapsible from '@/components/Collapsible/Collapsible'
 import { Pillow } from '@/components/Pillow/Pillow'
 import { Skeleton } from '@/components/Skeleton'
 import { ProductDetails } from './ProductDetails'
