+++ apps/store/src/components/ShopBreakdown/DiscountField.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled'
 import { useTranslation } from 'next-i18next'
 import { useState } from 'react'
 import { Text, theme } from 'ui'
-import * as Collapsible from '@/components/Collapsible'
+import Collapsible from '@/components/Collapsible/Collapsible'
 import { Switch } from '@/components/Switch'
 import { AddCampaignForm } from './AddCampaignForm'
 import { AddedCampaignForm } from './AddedCampaignForm'
