+++ apps/store/src/features/widget/ProductItem.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'next-i18next'
 import React, { useState, useMemo, type ReactNode } from 'react'
 import { CrossIconSmall, LockIcon, Text, theme } from 'ui'
 import { CancellationForm } from '@/components/Cancellation/CancellationForm'
-import * as Collapsible from '@/components/Collapsible'
+import Collapsible from '@/components/Collapsible/Collapsible'
 import { InputDay } from '@/components/InputDay/InputDay'
 import { Pillow } from '@/components/Pillow/Pillow'
 import { ProductDetails } from '@/components/ProductItem/ProductDetails'
